Authorization With Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -107,10 +107,16 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        if ($thread->user_id != auth()->id()) {
+            abort(403, 'not permitted');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
-            return response([]. 204);
+            return response([], 204);
         }
 
         return redirect('/threads');
